requirements.txt
# Core dependencies
gitpython>=3.1.0
colorama>=0.4.4

# Development dependencies (optional)
# Uncomment for development
# pytest>=7.0.0
# pytest-cov>=3.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950